requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
binance==0.3
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.0
captcha==0.5.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
dateparser==1.1.8
debugpy==1.6.6
decorator==5.1.1
discord.py==2.3.2
distlib==0.3.6
executing==1.2.0
filelock==3.9.0
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==4.0.0
forex-python==1.8
frozenlist==1.3.3
future==0.18.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gspread==5.9.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pandas==2.0.1
parso==0.8.3
paux==1.0.11
pbinance==1.0.6
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
protobuf==4.23.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.17
Pygments==2.14.0
pyparsing==3.0.9
python-binance==1.0.17
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.0
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
wcwidth==0.2.6
websocket-client==1.7.0
websockets==11.0.2
Werkzeug==2.3.3
yarl==1.9.2
zipp==3.14.0
